A few package org fixes + add ClusterDecoder interface
