+++ cmd/mathieson.go
@@ -1,4 +1,4 @@
-package cmd
+package main
 
 import (
 	"github.com/aphecetche/galo/mathieson"
