+++ cmd/plot.go
@@ -1,4 +1,4 @@
-package cmd
+package main
 
 import (
 	"os"
