+++ yaml/yaml_test.go
@@ -1,4 +1,4 @@
-package convert
+package yaml
 
 import (
 	"fmt"
